+++ pypt-extras/Qt/resources.py
@@ -1,14 +1,14 @@
 # -*- encoding: utf-8 -*-
 # Tara v0.1.0
 # Tara Python Wrapper
-# Copyright © 2020, H. Sinan Alioglu.
+# Copyright © 2020, Sinan Alioglu.
 # See /LICENSE for licensing information.
 # This file was adapted from Chris Warrick’s Python Project Template.
 
 """
 Adapt Qt resources to Python version.
 
-:Copyright: © 2020, H. Sinan Alioglu.
+:Copyright: © 2020, Sinan Alioglu.
 :License: BSD (see /LICENSE).
 """