+++ tara/template.py
@@ -1,13 +1,13 @@
 # -*- encoding: utf-8 -*-
 # Tara v0.1.0
 # Tara Python Wrapper
-# Copyright © 2020, H. Sinan Alioglu.
+# Copyright © 2020, Sinan Alioglu.
 # See /LICENSE for licensing information.
 
 """
 INSERT MODULE DESCRIPTION HERE.
 
-:Copyright: © 2020, H. Sinan Alioglu.
+:Copyright: © 2020, Sinan Alioglu.
 :License: BSD (see /LICENSE).
 """